+++ src/ResourceManager/Automation/Commands.Automation/Common/RequiresExtensions.cs
@@ -25,6 +25,8 @@ namespace Microsoft.Azure.Commands.Automation.Common
 
         private const string AccountNameValidator = "^[A-Za-z][-A-Za-z0-9]{4,48}[A-Za-z0-9]$";
 
+        private const string NodeConfigurationValidator = "^[^\\r\\n\\f<>*%&:?+/\\\\]{0,128}$";
+
         #endregion
 
         #region Public Methods and Operators
@@ -51,6 +53,33 @@ namespace Microsoft.Azure.Commands.Automation.Common
                     string.Format(
                         CultureInfo.CurrentCulture,
                         Resources.AutomationAccountNotFound));
+            }
+
+            return argument;
+        }
+
+        /// <summary>
+        /// Validates that the provided automation account name is valid.
+        /// </summary>
+        /// <param name="argument">
+        /// The argument.
+        /// </param>
+        /// <returns>
+        /// The <see cref="Requires.ArgumentRequirements{T}"/>.
+        /// </returns>
+        public static Requires.ArgumentRequirements<string> ValidNodeConfigurationName(this Requires.ArgumentRequirements<string> argument)
+        {
+            Requires.Argument(argument.Name, argument.Value).NotNull();
+
+            string stringValue = argument.Value;
+
+            if (!new Regex(NodeConfigurationValidator).IsMatch(stringValue))
+            {
+                // CDM TFS 665994 - we decided to display AutomationAccountNotFound even if the account name is invalid.
+                throw new ArgumentException(
+                    string.Format(
+                        CultureInfo.CurrentCulture,
+                        Resources.NodeConfigurationNameInvalid));
             }
 
             return argument;